fix spell miss
